requirements.txt
Pillow
open3d
numpy
scipy
opencv-python
shapely
matplotlib
# py360convert